Apollo: rootdir: Restore default swappiness parameters
In some regions, Android apps can be overly bloated, and setting the
memory limit to 60 can significantly increase memory pressure, especially
when physical memory is insufficient and not fully compressed and swapped
into ZRAM. This can slow down memory allocation and may even cause background
apps to be killed due to sudden increases in memory pressure.

This issue is more pronounced in devices with 6GB and 8GB of RAM.
Test: Using the same software (TikTok, WeChat, Bilibili, JD, Gaode Maps, Taobao),
we observed zram utilization, background persistence, and memory pressure.
A swappiness of 100 is significantly better than 60 in daily use,
and the system doesn't become sluggish (in terms of performance fluctuations).

Change-Id: Ib9c8173795e1f36f1845d5e89fb6d7678dc73da9